+++ packages/consts/CHANGELOG.md
@@ -3,6 +3,154 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0](https://github.com/apify/apify-shared-js/compare/@apify/consts@2.2.0...@apify/consts@2.3.0) (2022-07-27)
+
+
+### Features
+
+* Increasing BUILD_DEFAULT_MEMORY_MBYTES to 2 GB ([#321](https://github.com/apify/apify-shared-js/issues/321)) ([e9bdf1f](https://github.com/apify/apify-shared-js/commit/e9bdf1fb81fa282b9bd616c6aabc546b592bf810))
+
+
+
+
+
+# [2.2.0](https://github.com/apify/apify-shared-js/compare/@apify/consts@2.1.0...@apify/consts@2.2.0) (2022-07-26)
+
+
+### Features
+
+* add WEBHOOK_DISPATCH_STATUSES const ([#317](https://github.com/apify/apify-shared-js/issues/317)) ([0416c73](https://github.com/apify/apify-shared-js/commit/0416c73c95809a1b9c5b41e030a74fb981c51802))
+
+
+
+
+
+# [2.1.0](https://github.com/apify/apify-shared-js/compare/@apify/consts@2.0.0...@apify/consts@2.1.0) (2022-07-19)
+
+
+### Features
+
+* **consts:** add a new platform limit for max concurrent actor runs ([#318](https://github.com/apify/apify-shared-js/issues/318)) ([630de0b](https://github.com/apify/apify-shared-js/commit/630de0b2169a48cac1f5b4a83a0d055d6a3a2b5d))
+
+
+
+
+
+# [2.0.0](https://github.com/apify/apify-shared-js/compare/@apify/consts@1.11.0...@apify/consts@2.0.0) (2022-06-21)
+
+
+### Features
+
+* dual (native) ESM/CJS support all the packages ([#312](https://github.com/apify/apify-shared-js/issues/312)) ([daf882e](https://github.com/apify/apify-shared-js/commit/daf882ecdb3ff5b75975b92fc3528802a53bc736))
+
+
+### BREAKING CHANGES
+
+* All packages now have dual ESM/CJS build and require node 14+.
+
+
+
+
+
+# [1.11.0](https://github.com/apify/apify-shared-js/compare/@apify/consts@1.10.1...@apify/consts@1.11.0) (2022-06-02)
+
+
+### Features
+
+* **consts:** add string version of comma separated emails regex ([#308](https://github.com/apify/apify-shared-js/issues/308)) ([3ef91bb](https://github.com/apify/apify-shared-js/commit/3ef91bb24868f7246137a4587e859d52ebf022bb))
+
+
+
+
+
+## [1.10.1](https://github.com/apify/apify-shared-js/compare/@apify/consts@1.10.0...@apify/consts@1.10.1) (2022-05-31)
+
+
+### Bug Fixes
+
+* **consts:** remove g flag from regex ([#305](https://github.com/apify/apify-shared-js/issues/305)) ([50f92ca](https://github.com/apify/apify-shared-js/commit/50f92ca82d7e78b43b7879641dcc2f708ac96626))
+
+
+
+
+
+# [1.10.0](https://github.com/apify/apify-shared-js/compare/@apify/consts@1.9.0...@apify/consts@1.10.0) (2022-05-25)
+
+
+### Features
+
+* **consts:** adds regex for comma separated list of emails ([#302](https://github.com/apify/apify-shared-js/issues/302)) ([0f3a247](https://github.com/apify/apify-shared-js/commit/0f3a247c2ea6a22b65e07fec6ae02afb3eb4e10f))
+
+
+
+
+
+# [1.9.0](https://github.com/apify/apify-shared-js/compare/@apify/consts@1.8.0...@apify/consts@1.9.0) (2022-05-19)
+
+
+### Features
+
+* Adding build webhook types ([b630c46](https://github.com/apify/apify-shared-js/commit/b630c461e7cb8a89c401b304c4c8fa93b8c058b1))
+
+
+
+
+
+# [1.8.0](https://github.com/apify/apify-shared-js/compare/@apify/consts@1.7.0...@apify/consts@1.8.0) (2022-05-18)
+
+
+### Features
+
+* **consts:** add env var consts for actor build ID and build number ([#301](https://github.com/apify/apify-shared-js/issues/301)) ([b6fcb02](https://github.com/apify/apify-shared-js/commit/b6fcb02b9c1ff4dcfc83e0752768182139bed1ac))
+
+
+
+
+
+# [1.7.0](https://github.com/apify/apify-shared-js/compare/@apify/consts@1.6.0...@apify/consts@1.7.0) (2022-01-21)
+
+
+### Features
+
+* **consts:** remove unused consts ([#279](https://github.com/apify/apify-shared-js/issues/279)) ([4663d0f](https://github.com/apify/apify-shared-js/commit/4663d0f05e6157dfb5571f2ceb451db70d834987))
+
+
+
+
+
+# [1.6.0](https://github.com/apify/apify-shared-js/compare/@apify/consts@1.5.1...@apify/consts@1.6.0) (2022-01-21)
+
+
+### Features
+
+* **consts:** added deprecated base images still used by some actors ([#277](https://github.com/apify/apify-shared-js/issues/277)) ([e84e4b7](https://github.com/apify/apify-shared-js/commit/e84e4b7cc29a67f66d0bf1f6b1ecdf771c1d716a))
+
+
+
+
+
+## [1.5.1](https://github.com/apify/apify-shared-js/compare/@apify/consts@1.5.0...@apify/consts@1.5.1) (2022-01-11)
+
+
+### Bug Fixes
+
+* consecutive dots in email are not allowed ([#272](https://github.com/apify/apify-shared-js/issues/272)) ([25c4cbe](https://github.com/apify/apify-shared-js/commit/25c4cbe730ca9de7ac34f6fe368e63a58f4c0a67))
+
+
+
+
+
+# [1.5.0](https://github.com/apify/apify-shared-js/compare/@apify/consts@1.4.2...@apify/consts@1.5.0) (2021-12-14)
+
+
+### Features
+
+* Added new actor store categories ([#271](https://github.com/apify/apify-shared-js/issues/271)) ([8b50716](https://github.com/apify/apify-shared-js/commit/8b50716cc3301411d94ef01c711ff515d5d6b152))
+
+
+
+
+
 ## [1.4.2](https://github.com/apify/apify-shared-js/compare/@apify/consts@1.4.1...@apify/consts@1.4.2) (2021-11-15)
 
 **Note:** Version bump only for package @apify/consts