+++ packages/input_schema/CHANGELOG.md
@@ -3,6 +3,151 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.3](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@3.1.2...@apify/input_schema@3.1.3) (2022-08-29)
+
+
+### Bug Fixes
+
+* enforce unique items in "enum" input schema properties ([#329](https://github.com/apify/apify-shared-js/issues/329)) ([e2ff9a4](https://github.com/apify/apify-shared-js/commit/e2ff9a4fda6170c1c4f74c1c68928513bee444bb)), closes [#133](https://github.com/apify/apify-shared-js/issues/133)
+
+
+
+
+
+## [3.1.2](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@3.1.1...@apify/input_schema@3.1.2) (2022-07-27)
+
+**Note:** Version bump only for package @apify/input_schema
+
+
+
+
+
+## [3.1.1](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@3.1.0...@apify/input_schema@3.1.1) (2022-07-26)
+
+**Note:** Version bump only for package @apify/input_schema
+
+
+
+
+
+# [3.1.0](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@3.0.1...@apify/input_schema@3.1.0) (2022-07-25)
+
+
+### Features
+
+* **input_schema:** better error message for wrong enum value ([#320](https://github.com/apify/apify-shared-js/issues/320)) ([b4ce146](https://github.com/apify/apify-shared-js/commit/b4ce146b076789488245e33ca7f0c66922825bf4))
+
+
+
+
+
+## [3.0.1](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@3.0.0...@apify/input_schema@3.0.1) (2022-07-19)
+
+**Note:** Version bump only for package @apify/input_schema
+
+
+
+
+
+# [3.0.0](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@2.1.0...@apify/input_schema@3.0.0) (2022-06-21)
+
+
+### Features
+
+* dual (native) ESM/CJS support all the packages ([#312](https://github.com/apify/apify-shared-js/issues/312)) ([daf882e](https://github.com/apify/apify-shared-js/commit/daf882ecdb3ff5b75975b92fc3528802a53bc736))
+
+
+### BREAKING CHANGES
+
+* All packages now have dual ESM/CJS build and require node 14+.
+
+
+
+
+
+# [2.1.0](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@2.0.13...@apify/input_schema@2.1.0) (2022-06-10)
+
+
+### Features
+
+* Allow $schema prop in Input schema ([#310](https://github.com/apify/apify-shared-js/issues/310)) ([78385a6](https://github.com/apify/apify-shared-js/commit/78385a637e1e23e802a44b3dcc25e841131772ac))
+
+
+
+
+
+## [2.0.13](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@2.0.12...@apify/input_schema@2.0.13) (2022-06-02)
+
+**Note:** Version bump only for package @apify/input_schema
+
+
+
+
+
+## [2.0.12](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@2.0.11...@apify/input_schema@2.0.12) (2022-05-31)
+
+**Note:** Version bump only for package @apify/input_schema
+
+
+
+
+
+## [2.0.11](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@2.0.10...@apify/input_schema@2.0.11) (2022-05-25)
+
+**Note:** Version bump only for package @apify/input_schema
+
+
+
+
+
+## [2.0.10](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@2.0.9...@apify/input_schema@2.0.10) (2022-05-19)
+
+**Note:** Version bump only for package @apify/input_schema
+
+
+
+
+
+## [2.0.9](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@2.0.8...@apify/input_schema@2.0.9) (2022-05-18)
+
+**Note:** Version bump only for package @apify/input_schema
+
+
+
+
+
+## [2.0.8](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@2.0.7...@apify/input_schema@2.0.8) (2022-01-21)
+
+**Note:** Version bump only for package @apify/input_schema
+
+
+
+
+
+## [2.0.7](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@2.0.6...@apify/input_schema@2.0.7) (2022-01-21)
+
+**Note:** Version bump only for package @apify/input_schema
+
+
+
+
+
+## [2.0.6](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@2.0.5...@apify/input_schema@2.0.6) (2022-01-11)
+
+**Note:** Version bump only for package @apify/input_schema
+
+
+
+
+
+## [2.0.5](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@2.0.4...@apify/input_schema@2.0.5) (2021-12-14)
+
+**Note:** Version bump only for package @apify/input_schema
+
+
+
+
+
 ## [2.0.4](https://github.com/apify/apify-shared-js/compare/@apify/input_schema@2.0.3...@apify/input_schema@2.0.4) (2021-11-15)
 
 **Note:** Version bump only for package @apify/input_schema